backend/pyproject.toml
[project]
name = "ortho-backend"
version = "0.1.0"
description = "Agentic orthopedic backend powered by FastAPI, LangGraph, and MCP"
readme = "README.md"
requires-python = ">=3.11,<3.14"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-multipart>=0.0.12",
    "langchain>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-groq>=0.2.0",
    "langchain-community>=0.3.0",
    "langgraph>=0.2.0",
    "langgraph-checkpoint>=2.0.0",
    "langsmith>=0.2.0",
    "mcp>=1.0.0",
    "groq>=0.11.0",
    "torch>=2.1.0",
    "torchvision>=0.16.0",
    "ultralytics>=8.3.0",
    "pillow>=10.0.0",
    "numpy>=1.26.0",
    "opencv-python-headless>=4.9.0",
    "reportlab>=4.2.0",
    "cloudinary>=1.41.0",
    "aiofiles>=24.1.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "python-dotenv>=1.0.0",
    "loguru>=0.7.0",
    "httpx>=0.27.0",
    "python-jose>=3.3.0",
    "passlib>=1.7.4",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["api*", "core*", "graph*", "agents*", "services*", "tools*"]
exclude = ["models*", "storage*", "mcp*"]
